refactor to ensure sellers only delete prodcuts they uploaded

+++ controllers/products.js
@@ -62,6 +62,7 @@ const updateProduct = async (req, res) => {
 
   const {
     body: { name, category, price, description },
+    user: { userId },
     params: { id: productId },
   } = req;
 
@@ -69,13 +70,13 @@ const updateProduct = async (req, res) => {
     throw new BadRequestError('product name, category, price, or description fields cannot be empty')
   }
 
-  const product = await Product.findByIdAndUpdate({ _id: productId }, req.body, {
+  const product = await Product.findByIdAndUpdate({ _id: productId, createdBy: userId }, req.body, {
     new: true,
     runValidators: true
   });
 
   if (!product) {
-    throw new NotFoundError(`No product with id ${productId}`)
+    throw new NotFoundError(`No product with id ${productId}`);
   }
 
   res.status(StatusCodes.OK).json({ product })
@@ -85,11 +86,12 @@ const updateProduct = async (req, res) => {
 const deleteProduct = async (req, res) => {
 
   const {
+    user: { userId },
     params: { id: productId }
   } = req;
 
   // Note this can delete another product
-  const product = await Product.findOneAndDelete({ _id: productId })
+  const product = await Product.findOneAndDelete({ _id: productId, createdBy: userId })
 
   if (!product) {
     throw new NotFoundError(`No product with id : ${productId}`)
